Make db.close work properly
Also includes a hack of sorts to stop the http agent from using more than one connection. Due to the nature of this library it needs a bit of control over the connection itself and that can only be done when there's a single connection. The absense of multiple connections doesn't appear to be an issue. Its still possible to individually store over 2000 items in well under half a second.